requirements.txt
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
albumentations>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
shapely>=1.8.0
PyYAML>=6.0
tqdm>=4.62.0
Pillow>=8.3.0
